add jsr validation, change big decimal to double

+++ gamestore/src/main/java/com/company/gamestore/repositories/ProcessingFeeRepository.java
@@ -9,5 +9,5 @@ import java.math.BigDecimal;
 @Repository
 public interface ProcessingFeeRepository extends JpaRepository<Fee, String> {
 
-    BigDecimal findFeeByProductType(String itemType);
+    Double findFeeByProductType(String itemType);
 }
